lab12

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
+
 
 class ArticleController extends Controller
 {
@@ -22,6 +24,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create',[self::class]);
         return view('article.create');
     }
 
@@ -30,6 +33,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create',[self::class]);
         $request->validate([
             'date' => 'required',
             'title' => 'required|min:6',
@@ -51,7 +55,7 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        $comments = Comment::where('article_id', $article->id)->latest()->get();
+        $comments = Comment::where('article_id', $article->id)->where('accept', true)->latest()->get();
         return view('article.show', ['article'=>$article, 'comments'=>$comments]);
     }
 
@@ -60,6 +64,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('create',[self::class]);
         return view('article.edit', ['article'=>$article]);
     }
 
@@ -68,6 +73,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('create',[self::class]);
         $request->validate([
             'date' => 'required',
             'title' => 'required|min:6',
@@ -87,6 +93,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('create',[self::class]);
         $article->delete();
         return redirect()->route('article.index');
     }
